Move graph node class to graph_node.
